Compatibility with ChiliProject. Fixed display of code review popup dialog.
Slightly modified title of code review view. It now includes the tracker name and review status. This mimics the issue view of both Redmine and ChiliProject.

Open issues when used with ChiliProject:

1) Links to comments are incorrect.
2) Link to quote comment does not work.

Note that both of these features are not available when using the plugin with Redmine. Thus, removing them from the ChiliProject version might be a valid solution. However, the quoting functionality would be a nice to have.